# openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "RTCAS API",
    "version": "1.0.0",
    "description": "RTCAS Backend API Definition"
  },
  "servers": [
    { "url": "http://localhost:3000" },
    { "url": "https://api.rtcas.in.th" }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "security": [{ "bearerAuth": [] }],
  "paths": {
    "/api/admin/users": {
      "get": {
        "summary": "Get all users (admin)",
        "tags": ["Admin Users"],
        "security": [{ "bearerAuth": [] }],
        "responses": { "200": { "description": "List of users" } }
      },
      "post": {
        "summary": "Create user (admin)",
        "tags": ["Admin Users"],
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": false,
          "content": {
            "multipart/form-data": { "schema": { "type": "object" } }
          }
        },
        "responses": { "201": { "description": "User created" } }
      }
    },
    "/api/admin/users/personal-search": {
      "get": {
        "summary": "Search user personal information",
        "tags": ["Admin Users"],
        "security": [{ "bearerAuth": [] }],
        "responses": { "200": { "description": "Search result" } }
      }
    },
    "/api/admin/users/students": {
      "get": {
        "summary": "Get all students",
        "tags": ["Admin Users"],
        "security": [{ "bearerAuth": [] }]
      }
    },
    "/api/admin/users/teachers": {
      "get": {
        "summary": "Get all teachers",
        "tags": ["Admin Users"],
        "security": [{ "bearerAuth": [] }]
      }
    },
    "/api/admin/users/{id}": {
      "get": {
        "summary": "Get user by id",
        "tags": ["Admin Users"],
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": { "type": "integer" }
          }
        ]
      },
      "put": {
        "summary": "Update user",
        "tags": ["Admin Users"],
        "security": [{ "bearerAuth": [] }]
      }
    },
    "/api/admin/users/{id}/avatar": {
      "post": {
        "summary": "Upload user avatar",
        "tags": ["Admin Users"],
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "avatar": { "type": "string", "format": "binary" }
                }
              }
            }
          }
        }
      }
    },
    "/api/admin/users/deactivate/{id}": {
      "delete": {
        "summary": "Deactivate user",
        "tags": ["Admin Users"],
        "security": [{ "bearerAuth": [] }]
      }
    },
    "/api/admin/users/activate/{id}": {
      "patch": {
        "summary": "Activate user",
        "tags": ["Admin Users"],
        "security": [{ "bearerAuth": [] }]
      }
    },
    "/api/admin/training-reports": {
      "get": {
        "summary": "Training report dashboard",
        "tags": ["Admin Training"],
        "security": [{ "bearerAuth": [] }]
      }
    },
    "/api/admin/teacher-leaves/summary": {
      "get": {
        "summary": "Teacher leave summary",
        "tags": ["Admin Leaves"],
        "security": [{ "bearerAuth": [] }]
      }
    },
    "/api/admin/teacher-leaves": {
      "get": {
        "summary": "List teacher leaves",
        "tags": ["Admin Leaves"],
        "security": [{ "bearerAuth": [] }]
      }
    },
    "/api/admin/teacher-leaves/{id}/status": {
      "patch": {
        "summary": "Update teacher leave status",
        "tags": ["Admin Leaves"],
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": { "type": "integer" }
          }
        ]
      }
    },
    "/api/admin/teacher-leaves/current": {
      "get": {
        "summary": "List current teacher leaves",
        "tags": ["Admin Leaves"],
        "security": [{ "bearerAuth": [] }]
      }
    },
    "/api/admin/teaching-schedules": {
      "post": {
        "summary": "Create teaching schedule",
        "tags": ["Admin Schedule"],
        "security": [{ "bearerAuth": [] }]
      }
    },
...
    "/api/admin/teaching-schedules/{id}": {
      "put": {
        "summary": "Update teaching schedule",
        "tags": ["Admin Schedule"],
        "security": [{ "bearerAuth": [] }]
      },
      "delete": {
        "summary": "Delete teaching schedule",
        "tags": ["Admin Schedule"],
        "security": [{ "bearerAuth": [] }]
      }
    },
    "/api/admin/tasks": {
      "post": {
        "summary": "Create task",
        "tags": ["Admin Tasks"],
        "security": [{ "bearerAuth": [] }]
      },
      "get": {
        "summary": "List tasks",
        "tags": ["Admin Tasks"],
        "security": [{ "bearerAuth": [] }]
      }
    },
    "/api/admin/tasks/{id}": {
      "patch": {
        "summary": "Update task status",
        "tags": ["Admin Tasks"],
        "security": [{ "bearerAuth": [] }]
      },
      "delete": {
        "summary": "Delete task",
        "tags": ["Admin Tasks"],
        "security": [{ "bearerAuth": [] }]
      }
    },
...
    "/api/discipline-assessments/import": {
      "post": {
        "summary": "Import discipline assessments from Excel",
        "tags": ["Discipline Assessments"],
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "required": ["file"],
                "properties": {
                  "file": { "type": "string", "format": "binary" }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Import success" },
          "400": { "description": "Invalid file or data" },
          "401": { "description": "Unauthorized" },
          "403": { "description": "Forbidden" }
        }
      }
    },
    "/api/discipline-assessments": {
      "get": {
        "summary": "List discipline assessments",
        "tags": ["Discipline Assessments"],
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "in": "query",
            "name": "page",
            "required": false,
            "schema": { "type": "integer" }
          },
          {
            "in": "query",
            "name": "pageSize",
            "required": false,
            "schema": { "type": "integer" }
          }
        ],
        "responses": {
          "200": { "description": "List returned" },
          "401": { "description": "Unauthorized" },
          "403": { "description": "Forbidden" }
        }
      },
      "delete": {
        "summary": "Delete all discipline assessments",
        "tags": ["Discipline Assessments"],
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": { "description": "Deleted all" },
          "401": { "description": "Unauthorized" },
          "403": { "description": "Forbidden" }
        }
      }
    },
    "/api/discipline-assessments/{id}": {
      "delete": {
        "summary": "Delete discipline assessment by id",
        "tags": ["Discipline Assessments"],
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": { "type": "integer" }
          }
        ],
        "responses": {
          "200": { "description": "Deleted" },
          "401": { "description": "Unauthorized" },
          "403": { "description": "Forbidden" },
          "404": { "description": "Not found" }
        }
      }
    },
...
  },
...
}